--- pyparse/pyfront/namespace.py
@@ -1,1 +1,0 @@
-from ..pyfront import transform as transform
